Fix bug in dedup database code.
The dedup of the OSM and REPD tables both created a temporary view called parts, and another one
called clusters. I assumed that these temporary views would be deleted "after the current
session". However, it turns out that "the current session" lasted at least the db creation time (ie,
the duration of the the psql command file) and so there was a clash.

These tables have been renamed to include the prefix `osm_` or `repd_` respectively.